requirements.txt
scipion-pyworkflow
scipion-em
mrcfile
tensorboardX
scipy
